Move enemy spawner function to Game class
Move enemySpawn to Game class and rename to spawnEnemy and delete
enemySpawn.js
