+++ src/game.js
@@ -1,5 +1,6 @@
 import gameBoard from './gameBoard';
 import castle from './castle';
+import Enemy from './enemy';
 
 class Game {
     constructor() {
@@ -14,6 +15,22 @@ class Game {
 
     draw() {
         this.enemies.forEach((enemy) => enemy.draw());
+    }
+
+    spawnEnemy() {
+        const enemiesCount = 3;
+        const yPos = [62.5, 175, 287.5];
+        const gameBoardElement = gameBoard.element;
+
+        for (let i = 0; i < enemiesCount; i += 1) {
+            const enemy = new Enemy(0, yPos[i]);
+            this.enemies.push(enemy);
+        }
+
+        this.enemies.forEach((enemy) => {
+            gameBoardElement.appendChild(enemy.element);
+            enemy.draw();
+        });
     }
 }
